requirements.txt
Flask
Flask-SocketIO
websockets
python-dotenv
Flask-Session

